+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Drop support for Plone 4.0 and 4.1.
   [hvelarde]
 
+- Folder were not selectable as root object [keul]
 - Remove also browserlayer when running unistall profile. [keul]
 - Uninstall profiles was not run on uninstall [keul]
 - Removed useless paster generated code (there's no Archetypes contents there).